+++ src/components/mergeRectangles.js
